Remove Mongo.
