+++ .travis.yml
@@ -1,9 +1,5 @@
 language: node_js
 node_js:
   - "4.2"
-services:
-  - mongodb
-before_script:
-  - mongo mydb_test --eval 'db.addUser("travis", "test");'
 script:
   - travis_wait npm run test
